feat: refactor with Vite for simpler usage
