+++ .gitignore
@@ -1,1 +1,2 @@
 .env
+node_modules
