+++ frontend/Dockerfile
@@ -2,12 +2,17 @@ FROM node:18-alpine
 
 WORKDIR /app
 
+# Copy package.json and package-lock.json first for better caching
 COPY package*.json ./
 
+# Install dependencies
 RUN npm install
 
+# Copy the rest of the application code
 COPY . .
 
+# Expose the port the app will run on
 EXPOSE 3000
 
-CMD ["npm", "run", "dev"]
+# Command to run the application
+CMD ["npm", "run", "dev", "--", "--host", "0.0.0.0"]
