+++ frontend/package.json
@@ -1,21 +1,25 @@
 {
     "name": "circlematch-frontend",
-    "version": "0.1.0",
     "private": true,
+    "version": "0.1.0",
+    "type": "module",
     "scripts": {
-      "dev": "next dev",
-      "build": "next build",
-      "start": "next start"
+      "dev": "vite",
+      "build": "vite build",
+      "preview": "vite preview"
     },
     "dependencies": {
-      "next": "14.0.4",
       "react": "18.2.0",
       "react-dom": "18.2.0",
-      "@react-oauth/google": "^0.8.0"
+      "@react-oauth/google": "^0.8.0",
+      "react-router-dom": "^6.21.1"
     },
     "devDependencies": {
       "@types/node": "20.10.5",
       "@types/react": "18.2.45",
-      "typescript": "5.3.3"
+      "@types/react-dom": "18.2.18",
+      "@vitejs/plugin-react": "^4.2.1",
+      "typescript": "5.3.3",
+      "vite": "^5.0.10"
     }
   }
