+++ frontend/src/App.tsx
@@ -1,0 +1,17 @@
+import { Routes, Route } from 'react-router-dom'
+import { GoogleOAuthProvider } from '@react-oauth/google'
+import HomePage from './pages/Home'
+import TeacherPage from './pages/Teacher'
+
+function App() {
+  return (
+    <GoogleOAuthProvider clientId={import.meta.env.VITE_GOOGLE_CLIENT_ID || ""}>
+      <Routes>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/teacher" element={<TeacherPage />} />
+      </Routes>
+    </GoogleOAuthProvider>
+  )
+}
+
+export default App
