+++ frontend/src/components/DebugDropdown.tsx
@@ -12,7 +12,7 @@ const DebugDropdown: React.FC<DebugDropdownProps> = ({
   const [showDropdown, setShowDropdown] = useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
 
   const handleMouseEnter = () => {
     // Clear any existing timeout to prevent the dropdown from closing
