+++ frontend/src/components/UserDropdown.tsx
@@ -15,7 +15,7 @@ const UserDropdown: React.FC<UserDropdownProps> = ({
   const [showDropdown, setShowDropdown] = useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
+  const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
   
   const handleMouseEnter = () => {
     // Clear any existing timeout to prevent the dropdown from closing
