+++ frontend/vite.config.js
@@ -1,0 +1,32 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import path from 'path'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
+  },
+  server: {
+    port: 3000,
+    host: true, // Allow connections from outside the container
+    watch: {
+      usePolling: true, // Needed for Docker volumes
+    },
+    proxy: {
+      '/api': {
+        target: 'http://backend:8000',
+        changeOrigin: true,
+        secure: false,
+      },
+    },
+  },
+  define: {
+    // Define environment variables 
+    'process.env.NEXT_PUBLIC_API_URL': JSON.stringify(process.env.VITE_API_URL || 'http://localhost:8000'),
+    'process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID': JSON.stringify(process.env.VITE_GOOGLE_CLIENT_ID || ''),
+  },
+})
